Switched to use condition_variable_any to allow use of util::VMutex.

+++ src/util/Mutex.cc
@@ -39,7 +39,7 @@ LOG_LOGGER _log = LOG_GET("lsst.qserv.util.Mutex");
 namespace lsst::qserv::util {
 
 mutex Mutex::_lockedIdMtx;
-set<unsigned int> Mutex::_lockedId;
+set<uint64_t> Mutex::_lockedId;
 
 void Lock::_lock() {
     if (!_context.empty()) {
